change float to double

+++ src/beatmap/BeatmapV2.java
@@ -2,7 +2,7 @@ package beatmap;
 
 public class BeatmapV2 {
     public static class Note {
-        public float _time;
+        public double _time;
         public int _lineIndex;
         public int _lineLayer;
         public int _type;
@@ -10,15 +10,15 @@ public class BeatmapV2 {
     }
     public static class Obstacle {
         public int type;
-        public float _time;
-        public float _duration;
+        public double _time;
+        public double _duration;
         public int _lineIndex;
         public int width;
     }
     public static class Event {
-        public float _time;
+        public double _time;
         public int _type;
-        public float _floatValue;
+        public double _floatValue;
     }
 
     public String _version;
